Replacing all jar classes with JBL
Excluding all the converted Java -> C# classes this is never going to
work since there are specific calls in the C++ code that is requiring
Java classes it seems. So moving to JBL for all classes except the main
Activity.
Adding the jars to the main project as well.
Changing namespaces to handle the new changes
App is running, all screens seems to be working except actuall making a
video call